Move the web server in a new crate
