+++ Cargo.toml
@@ -32,7 +32,7 @@ tree-sitter = "^0.17"
 pretty_assertions = "^0.6"
 
 [workspace]
-members = ["rust-code-analysis-cli"]
+members = ["rust-code-analysis-cli", "rust-code-analysis-web"]
 exclude = ["enums"]
 
 [profile.release]
