+++ rust-code-analysis-cli/Cargo.toml
@@ -11,11 +11,8 @@ license = "MPL-2.0"
 name = "rust-code-analysis-cli"
 
 [dependencies]
-actix-rt = "^1.0"
-actix-web = "^3.2"
 clap = "^2.33"
 crossbeam = "^0.8"
-futures = "^0.3"
 globset = "^0.4"
 num_cpus = "^1.13"
 regex = "^1.4"
@@ -26,6 +23,3 @@ serde_json = "^1.0"
 serde_yaml = "^0.8"
 toml = "^0.5"
 walkdir = "^2.2"
-
-[dev-dependencies]
-pretty_assertions = "^0.6"
