+++ rust-code-analysis-web/Cargo.toml
@@ -1,0 +1,24 @@
+[package]
+name = "rust-code-analysis-web"
+version = "0.0.18"
+authors = ["Calixte Denizet <cdenizet@mozilla.com>"]
+edition = "2018"
+keywords = ["metrics"]
+description = "Run a web service to compute and export code metrics"
+license = "MPL-2.0"
+
+[[bin]]
+name = "rust-code-analysis-web"
+
+[dependencies]
+actix-rt = "^1.0"
+actix-web = "^3.2"
+clap = "^2.33"
+futures = "^0.3"
+num_cpus = "^1.13"
+rust-code-analysis = { path = ".." }
+serde = "^1.0"
+serde_json = "^1.0"
+
+[dev-dependencies]
+pretty_assertions = "^0.6"
